+++ opentelemetry-sdk/benches/span_builder.rs
@@ -6,7 +6,7 @@ use opentelemetry::{
 use opentelemetry_sdk::error::OTelSdkResult;
 use opentelemetry_sdk::{
     trace as sdktrace,
-    trace::{SpanData, SpanExporter},
+    trace::{SpanBatch, SpanExporter},
 };
 #[cfg(all(not(target_os = "windows"), feature = "bench_profiling"))]
 use pprof::criterion::{Output, PProfProfiler};
@@ -65,7 +65,7 @@ fn not_sampled_provider() -> (sdktrace::SdkTracerProvider, sdktrace::SdkTracer) 
 struct NoopExporter;
 
 impl SpanExporter for NoopExporter {
-    async fn export(&self, _spans: Vec<SpanData>) -> OTelSdkResult {
+    async fn export(&self, _spans: SpanBatch<'_>) -> OTelSdkResult {
         Ok(())
     }
 }